requirements.txt
PySide6
pydicom
numpy
pillow
segmentation_models_pytorch
opencv-python
torch
torchvision
pandas
openpyxl
